Move 'RunSyncQueryJob' from 'bigquery.job' -> 'bigquery.query'.
Move shared helpers to 'bigquery._helpers'.
